move App.js from `components` to `src/index.js`
